### 2016-06-17 #29 @whedon
Submission: Git-RDM: A research data management plugin for the Git version control system
**Submitting author:** @ctjacobs (<a href="http://orcid.org/0000-0002-0034-4650">Christian T. Jacobs</a>)
**Repository:** <a href="https://github.com/ctjacobs/git-rdm" target ="_blank">https://github.com/ctjacobs/git-rdm</a>
**Version:** v1.0.0
**Archive:** <a href="https://dx.doi.org/10.6084/m9.figshare.3443750.v1" target ="_blank">10.6084/m9.figshare.3443750.v1</a>

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.6084/m9.figshare.3439742.v1">software archive</a> resolve?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
