### 2016-06-17 #29 @desilinguist
:hand: I am reviewing this
On Thu, Jun 16, 2016 at 10:32 PM Arfon Smith <notifications@github.com>
wrote:

> / cc @openjournals/joss-reviewers
> <https://github.com/orgs/openjournals/teams/joss-reviewers> - would
> anyone be willing to review this submission?
>
> If you would like to review this submission then please comment on this
> thread so that others know you're doing a review (so as not to duplicate
> effort). Something as simple as :hand: I am reviewing this will suffice.
>
> *Reviewer instructions*
>
>    - Please work through the checklist at the start of this issue.
>    - If you need any further guidance/clarification take a look at the
>    reviewer guidelines here
>    http://joss.theoj.org/about#reviewer_guidelines
>    - Please make a publication recommendation at the end of your review
>
> Any questions, please ask for help by commenting on this issue! 🚀
>
