### 2016-06-17 #29 @jsta
No, setup.py has a version number of 1.0 while the github release is 1.0.0

- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.6084/m9.figshare.3439742.v1">software archive</a> resolve?

I am concerned about the instruction to install as root with sudo although I understand that this may be necessary in order for system-wide functionality. Also, I had to do a fair amount of fighting with PATH variables because I am a conda user.

Verified functionality for all items!

I would really prefer a more automated package installation solution like integration with conda.

Overall, I think `git-rdm` works great! I wonder if you could include instructions for viewing and verifying the `.rdm/publications.db` file.
- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?

There are no automated tests. However, I think the example usage in README.md is sufficient.
