### 2016-06-17 #29 @alex-konovalov
Similar thing happened to me with another paper last week. I was away from browser and was only able to reply by email, but my comment did not appear under that paper so it was reviewed by another person (thanks!).

Could it be a more general problem?
