### 2016-06-17 #29 @jsta
@desilinguist I saw your reply. However, I neglected to do a thorough check of other submissions or I would have seen that it was the norm to only have one reviewer per submission. I just assumed that multiple people reviewed each submission. My apologies!
