### 2016-06-17 #29 @ctjacobs
Thank you all so much for your interest in this, and for the extremely fast review! I'll start addressing your comments within the next few days.
