### 2016-06-17 #29 @jsta
I realized that my comment about `.rdm/publications.db` in the **Example Usage** section is solved by:
> `$ git rdm ls`

However, I ended up verifying the contents of the file separately using:
> `$ sqlite3 .rdm/publications.db '.header on' '.mode csv' '.once test.csv' 'SELECT * FROM publications'`

In summary, I can see some room for improvement in the way the output of `$ git rdm ls` is formatted.
