### 2016-06-18 f7b5bc1
Bump the version to v1.0.1, since we will be applying essentially a few patches.
