### 2016-06-18 b10bb23
Added a reference to the submitted JOSS paper.
