+++ README.md
@@ -177,6 +177,12 @@ git-rdm INFO: {
 ## License
 This software is released under the MIT license. See the file called `LICENSE` for more information.
 
+## Citing
+
+If you use Git-RDM during the course of your research, please consider citing the following paper:
+
+* C. T. Jacobs, A. Avdis (Under Review). Git-RDM: A research data management plugin for the Git version control system. *Journal of Open Source Software*.
+
 ## Contact
 Please send any questions or comments about Git-RDM via email to [Christian Jacobs](http://christianjacobs.uk) at <christian@christianjacobs.uk>.
 
