### 2016-06-19 4e943af
Add a note in the README about the --raw flag.
