### 2016-06-19 ff6f10c
Experimenting with conda.
