+++ conda/meta.yaml
@@ -1,0 +1,25 @@
+package:
+  name: git-rdm
+  version: "1.0.0"
+
+source:
+  git_url: https://github.com/ctjacobs/git-rdm.git
+  git_rev: 1.0.0
+
+build:
+  script: python setup.py install
+
+requirements:
+  build:
+    - python
+  run:
+    - python
+    - argparse
+    - gitpython
+    - sqlite3
+
+about:
+  home: https://github.com/ctjacobs/git-rdm
+  license: MIT
+  license_file: LICENSE
+  summary: a research data management plugin for the Git version control system
