### 2016-06-19 9c866b9
Added notes on installing Git-RDM via Conda.
