### 2016-06-19 41a39a0
Remove the conda directory.
