### 2016-06-19 8d3fd1c
Added a note about local user installations.
