### 2016-06-19 #29 @ctjacobs
I have addressed @jsta's comments in the [review branch](https://github.com/ctjacobs/git-rdm/tree/review) of the git-rdm repository. The main changes are:

* Added a new option `git rdm ls --raw` which outputs the raw contents of the `.rdm/publications.db` file for database verification purposes. Also added an example of its use in `README.md`.

* Bumped the version in `setup.py` to v1.0.1 (rather than deleting and re-releasing this revised code as v1.0.0); v1.0.1 will be released via the Git-RDM repository's [Releases page](https://github.com/ctjacobs/git-rdm/releases) (and archived again on Figshare) once everyone is happy with the changes I've made here.

* Packaged up Git-RDM and PyRDM (and the SWORD2 dependency), and uploaded it to [my anaconda.org package channel](https://anaconda.org/ctjacobs). Instructions for installing Git-RDM via Conda are now provided in `README.md`. Once again, the Conda package for Git-RDM will be updated to v1.0.1 once everyone is happy with the changes.
