### 2016-06-19 #29 @ctjacobs
For a full `diff`, please see https://github.com/ctjacobs/git-rdm/pull/2
