### 2016-06-19 #29 @jsta
Awesome! I was able to install from your conda channel and test the `git rdm ls --raw` command. I really like the improvements to `README.md` describing local installs. Everything is good-to-go for acceptance as far as I am concerned.
