### 2016-06-19 #29 @ctjacobs
Fantastic - many thanks for trying it out. I'll merge in the changes and update to v1.0.1.
