### 2016-06-19 0599bd8
Updated reference, following acceptance.
