### 2016-06-19 9a9ed67
Removed old tags from the bibtex references.
