+++ paper.bib
@@ -7,10 +7,7 @@
   Author                   = {Chacon, S. and Straub, B.},
   Publisher                = {Apress},
   Year                     = {2014},
-  Edition                  = {2nd},
-
-  Owner                    = {christian},
-  Timestamp                = {2016.06.15}
+  Edition                  = {2nd}
 }
 
 @Article{Jacobs_etal_2014,
@@ -21,10 +18,7 @@
   Number                   = {1},
   Pages                    = {e28},
   Volume                   = {2},
-
-  Doi                      = {10.5334/jors.bj},
-  Owner                    = {christian},
-  Timestamp                = {2015.07.06}
+  Doi                      = {10.5334/jors.bj}
 }
 
 
