### 2016-06-19 #29 @ctjacobs
All done. The archive for v1.0.1 is available at the following (different) DOI: https://dx.doi.org/10.6084/m9.figshare.3443750.v1

@arfon, is it possible to have the "Archive DOI" [here](http://joss.theoj.org/papers/d0641341031a50d63d61e6149768d19e) updated please?

The Conda package has also been updated to v1.0.1
