### 2016-06-20 #29 @arfon
Thanks for the excellent review @jsta!

@ctjacobs your paper DOI is http://dx.doi.org/10.21105/joss.00029 . The DOI isn't quite live yet (i.e. it doesn't resolve) but should start working in the next couple of hours once the Crossref queue processes the registration request. 🎉 🚀 💥
