### 2016-06-20 #29 @ctjacobs
Excellent. Thank you everyone!

@arfon, I would be happy to review future submissions for JOSS, so feel free to add me to @joss-reviewers.

Cheers!
